+++ src/index.js
@@ -2,12 +2,14 @@ import TmHelloWorld from "./components/tmHelloWorld"
 import TmField from "./components/tmField"
 import TmPageHeader from "./components/tmPageHeader"
 import TmNotifications from "./components/tmNotifications"
+import TmCountdown from "./components/tmCountdown"
 
 const TendermintComponents = {
   TmHelloWorld,
   TmField,
   TmPageHeader,
-  TmNotifications
+  TmNotifications,
+  TmCountdown
 }
 
 TendermintComponents.install = Vue =>
@@ -25,5 +27,9 @@ export { default as TmField } from "./components/tmField/tmField.vue"
 export {
   default as TmPageHeader
 } from "./components/tmPageHeader/tmPageHeader.vue"
+export {
   default as TmNotifications
 } from "./components/tmNotifications/tmNotifications.vue"
+export {
+  default as TmCountdown
+} from "./components/tmCountdown/tmCountdown.vue"